requirements.txt
aniso8601==9.0.1
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pdfkit==1.0.0
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.43
urllib3==1.26.12
Werkzeug==2.2.2
celery[redis] == 5.2.2
